Check exists appnba/data and create this directory

+++ download_gamelog.R
@@ -16,4 +16,9 @@ get_leaguegamelog <- function(Season=2020, ...){
 
 t <- get_leaguegamelog(2020)
 
-write.csv(t, 'appnba/data/gamelog.csv', row.names = FALSE)
+if (dir.exists('appnba/data')){
+  write.csv(t, 'appnba/data/gamelog.csv', row.names = FALSE)
+} else {
+  dir.create('appnba/data')
+  write.csv(t, 'appnba/data/gamelog.csv', row.names = FALSE)
+}
